backend/requirements.txt
celery==4.2
coverage==4.5
Django==2.1
djangorestframework
django-filter
djangorestframework-jwt
flake8==3.6
gunicorn==19.8
psycopg2==2.7
redis==2.10.5